ai_service/requirements.txt
flask
flask-cors
Pillow
numpy
tensorflow-cpu
google-genai
transformers
torch
